fix to use old are equal
